+++ packages/sqlite_async/lib/sqlite_async.dart
@@ -7,6 +7,7 @@ export 'src/common/abstract_open_factory.dart' hide InternalOpenFactory;
 export 'src/common/connection/sync_sqlite_connection.dart';
 export 'src/common/mutex.dart';
 export 'src/common/sqlite_database.dart' hide SqliteDatabaseImpl;
+export 'src/common/timeouts.dart' show AbortException;
 export 'src/sqlite_connection.dart';
 export 'src/sqlite_migrations.dart';
 export 'src/sqlite_options.dart';